pyproject.toml

[project]
name = "xlsx-streamer"
version = "0.1.0"
description = "Memory-efficient streaming XLSX to CSV converter for any data source."
authors = [
    { name = "Varun Prasad", email = "varunprasad717@gmail.com" }]
license = "MIT"
license-files = ["LICENSE*"]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "stream-unzip",
    "typer-slim>=0.21.1",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
s3 = ["boto3"]
http = ["httpx"]
all = ["boto3", "httpx"]


[project.urls]
Homepage = "https://github.com/varunprasad17/xlsx-streamer"
Repository = "https://github.com/varunprasad17/xlsx-streamer"

[project.scripts]
xlsx-streamer = "xlsx_streamer.cli:app"

[dependency-groups]
nox = [
    "nox-uv",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-randomly",
    "moto[s3]",
    "pytest-httpx",
    "boto3",
    "httpx",
    "openpyxl",
]
type_check = [
    "mypy",
]
lint = [
    "ruff",
]
docs = [
    "mkdocs-material",
    "mkdocs-htmlproofer-plugin",
    "mkdocstrings[python]",
    "mkdocs-api-autonav",
]
licenses = [
    "pip-licenses-cli",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

# For non-src directory projects, explicitly set top level model name:
#[tool.uv.build-backend]
#module-name = "my-app"
#module-root = ""

[tool.uv]
default-groups = "all"
# Non-package applications can uncomment the following line and delete [build-system] section to
# avoid being built or installed: https://docs.astral.sh/uv/reference/settings/#package
#package = false

[tool.mypy]
ignore_missing_imports = true
strict = true
enable_error_code = ["deprecated", "exhaustive-match", "explicit-override"]
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

[tool.ruff]
line-length = 99
src = ["src"]
# Ruff will automatically exclude all files listed in .gitignore as well as common temporary Python
# tool directories.
# To exclude additional folders, use extend-exclude.

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT", # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "RUF005",
    "RUF012",
]
unfixable = [
    # Disable removing unused imports by default and only enable within nox so editors don't delete
    # unused imports while the user is in the middle of editing a file on save.
    "F401",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning suppressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true

# To globally exclude additional code blocks from code coverage reporting, see:
# https://coverage.readthedocs.io/en/latest/excluding.html#advanced-exclusion
[tool.coverage.report]
exclude_also = [
    "case _ as unreachable",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
